# shared/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kubernetes Backup & Restore API
  description: |
    Comprehensive REST API for Kubernetes backup, restore, and disaster recovery operations.
    
    This API provides endpoints for:
    - Backup and restore operations
    - Disaster recovery scenario management
    - Cluster and resource validation
    - Integration bridge monitoring
    - Webhook management for GitOps workflows
    
    ## Authentication
    All API endpoints require appropriate authentication headers and role-based access control.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Check response headers for current limits.
    
    ## Circuit Breaker
    External service calls are protected by circuit breakers for enhanced resilience.
  version: "1.0.0"
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.backup-system.example.com
    description: Production server
paths:
...
  /api/v1/integration/events:
    post:
      summary: Publish integration event
      description: Publishes an event to the integration event bus
      operationId: publishEvent
      tags:
        - Integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationEvent'
      responses:
        '200':
          description: Event published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid event format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /webhook/backup:
    post:
      summary: Backup webhook
      description: Receives backup completion notifications
      operationId: handleBackupWebhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          description: Invalid webhook payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
  /webhook/gitops:
    post:
      summary: GitOps webhook
      description: Receives GitOps operation notifications
      operationId: handleGitOpsWebhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
  /webhook/restore:
    post:
      summary: Restore webhook
      description: Receives restore operation notifications
      operationId: handleRestoreWebhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'

  # Additional Webhook Registration Endpoints
  /register/backup:
    post:
      summary: Register backup webhook
      description: Registers a webhook endpoint for backup notifications
      operationId: registerBackupWebhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRegistration'
      responses:
        '200':
          description: Webhook registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
  /register/gitops:
    post:
      summary: Register GitOps webhook
      description: Registers a webhook endpoint for GitOps notifications
      operationId: registerGitOpsWebhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRegistration'
      responses:
        '200':
          description: Webhook registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
      required:
        - success
        - timestamp
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Operation failed"
        error:
          $ref: '#/components/schemas/APIError'
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
      required:
        - success
        - timestamp
    APIError:
      type: object
      properties:
        code:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Request validation failed"
        details:
          type: object
          additionalProperties: true
      required:
        - code
        - message

    # Health and Status Types
...
    IntegrationEvent:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
        event_type:
          type: string
          example: "backup_completed"
        source:
          type: string
          example: "backup-service"
        target:
          type: string
          example: "gitops-service"
        timestamp:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        priority:
          type: string
          enum: [low, normal, high, critical]
          example: "normal"
        dependencies:
          type: array
          items:
            type: string
      required:
        - event_type
        - source
        - timestamp
        - data
...
    WebhookRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          example: "backup_completed"
        source:
          type: string
          example: "backup-service"
        timestamp:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
        signature:
          type: string
          description: HMAC signature for webhook verification
          example: "sha256=abc123..."
      required:
        - id
        - type
        - source
        - timestamp
        - data
    WebhookResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Webhook processed successfully"
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        data:
          type: object
          additionalProperties: true
      required:
        - success
        - message
        - timestamp
    WebhookRegistration:
      type: object
      properties:
        endpoint_url:
          type: string
          format: uri
          example: "https://api.example.com/webhooks/backup"
        event_types:
          type: array
          items:
            type: string
          example: ["backup_completed", "backup_failed"]
        secret:
          type: string
          description: Secret for HMAC signature verification
        active:
          type: boolean
          default: true
        retry_policy:
          $ref: '#/components/schemas/RetryPolicy'
        headers:
          type: object
          additionalProperties:
            type: string
          description: Additional headers to include in webhook requests
      required:
        - endpoint_url
        - event_types
    RetryPolicy:
      type: object
      properties:
        max_attempts:
          type: integer
          minimum: 1
          maximum: 10
          default: 3
        initial_delay:
          type: string
          example: "1s"
        max_delay:
          type: string
          example: "5m"
        multiplier:
          type: number
          format: float
          minimum: 1.0
          default: 2.0
      required:
        - max_attempts
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    
    BasicAuth:
      type: http
      scheme: basic
      description: Basic HTTP authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - BasicAuth: []
...